refactoring thumbnail
